feat: support close, quit req task, sit test success

+++ test/sit/clean.sh
@@ -1,16 +1,35 @@
 #!/usr/bin/bash
 function clean_local(){
-    sudo rm -rf veth0
-    sudo rm -rf veth1
+    if [ -d "veth0" ]; then
+         sudo rm -rf veth0
+    fi
+    if [ -d "veth1" ]; then
+         sudo rm -rf veth1
+    fi
 }
 
 function clean_net() {
-    sudo ip netns del ns0
-    sudo ip netns del ns1
-    sudo ip link set dev vbridge down
-    sudo brctl delbr vbridge
-    sudo ip link delete veth-0-br
-    sudo ip link delete veth-1-br
+    if [[ $(ip netns | grep ns0) == ns0 ]]; then
+        sudo ip netns del ns0
+    fi
+
+    if [[ $(ip netns | grep ns1) == ns0 ]]; then
+        sudo ip netns del ns1
+    fi
+    
+    if [[ $(ip link | grep "vbridge:") == ns0 ]]; then
+        sudo ip link set dev vbridge down
+        sudo brctl delbr vbridge
+    fi
+
+    if [[ $(ip link | grep veth-0-br) == ns0 ]]; then
+         sudo ip link delete veth-0-br
+    fi
+
+    if [[ $(ip link | grep veth-1-br) == ns0 ]]; then
+         sudo ip link delete veth-1-br
+    fi
+    
 }
 function clean() {
     clean_local
